+++ src/router/routes.js
@@ -8,5 +8,10 @@ export default [
   {
     path: '/team',
     component: () => import('@/views/TeamList')
+  },
+  {
+    path: '/team/:id',
+    component: () => import('@/views/team/TeacherShow'),
+    props: true
   }
 ];