+++ src/views/team/TeacherShow.vue
@@ -1,0 +1,12 @@
+<template>
+  <b-container>
+    <p>Detailed info of Teacher ID: {{ id }} component Here</p>
+  </b-container>
+</template>
+
+<script>
+export default {
+  name: 'TeacherShow',
+  props: ['id']
+};
+</script>