+++ src/views/HomeView.vue
@@ -4,6 +4,7 @@
     <lab-intro />
     <teachers-brief />
     <lab-video />
+    <sponsors-list />
   </div>
 </template>
 
@@ -12,6 +13,7 @@ import NewsCarousel from '@/components/home/NewsCarousel';
 import LabIntro from '@/components/home/LabIntro';
 import TeachersBrief from '@/components/home/TeachersBrief';
 import LabVideo from '@/components/home/LabVideo';
+import SponsorsList from '@/components/home/SponsorsList';
 
 export default {
   name: 'HomeView',
@@ -19,7 +21,8 @@ export default {
     NewsCarousel,
     LabIntro,
     TeachersBrief,
-    LabVideo
+    LabVideo,
+    SponsorsList
   }
 };
 </script>